Move some codes to the parse_cp2k module
